+++ src/GraphQl/Resolver/Factory/ItemResolverFactory.php
@@ -21,7 +21,6 @@ use ApiPlatform\Core\GraphQl\Resolver\Stage\SerializeStageInterface;
 use ApiPlatform\Core\Metadata\Resource\Factory\ResourceMetadataFactoryInterface;
 use ApiPlatform\Core\Util\ClassInfoTrait;
 use ApiPlatform\Core\Util\CloneTrait;
-use GraphQL\Error\Error;
 use GraphQL\Type\Definition\ResolveInfo;
 use Psr\Container\ContainerInterface;
 
@@ -72,7 +71,7 @@ final class ItemResolverFactory implements ResolverFactoryInterface
                 throw new \LogicException('Item from read stage should be a nullable object.');
             }
 
-            $resourceClass = $this->getResourceClass($item, $resourceClass, $info);
+            $resourceClass = $this->getResourceClass($item, $resourceClass);
             $resourceMetadata = $this->resourceMetadataFactory->create($resourceClass);
 
             $queryResolverId = $resourceMetadata->getGraphqlAttribute($operationName, 'item_query');
@@ -80,7 +79,7 @@ final class ItemResolverFactory implements ResolverFactoryInterface
                 /** @var QueryItemResolverInterface $queryResolver */
                 $queryResolver = $this->queryResolverLocator->get($queryResolverId);
                 $item = $queryResolver($item, $resolverContext);
-                $resourceClass = $this->getResourceClass($item, $resourceClass, $info, sprintf('Custom query resolver "%s"', $queryResolverId).' has to return an item of class %s but returned an item of class %s.');
+                $resourceClass = $this->getResourceClass($item, $resourceClass, sprintf('Custom query resolver "%s"', $queryResolverId).' has to return an item of class %s but returned an item of class %s.');
             }
 
             ($this->securityStage)($resourceClass, $operationName, $resolverContext + [
@@ -102,13 +101,13 @@ final class ItemResolverFactory implements ResolverFactoryInterface
     /**
      * @param object|null $item
      *
-     * @throws Error
+     * @throws \UnexpectedValueException
      */
-    private function getResourceClass($item, ?string $resourceClass, ResolveInfo $info, string $errorMessage = 'Resolver only handles items of class %s but retrieved item is of class %s.'): string
+    private function getResourceClass($item, ?string $resourceClass, string $errorMessage = 'Resolver only handles items of class %s but retrieved item is of class %s.'): string
     {
         if (null === $item) {
             if (null === $resourceClass) {
-                throw Error::createLocatedError('Resource class cannot be determined.', $info->fieldNodes, $info->path);
+                throw new \UnexpectedValueException('Resource class cannot be determined.');
             }
 
             return $resourceClass;
@@ -121,7 +120,7 @@ final class ItemResolverFactory implements ResolverFactoryInterface
         }
 
         if ($resourceClass !== $itemClass) {
-            throw Error::createLocatedError(sprintf($errorMessage, (new \ReflectionClass($resourceClass))->getShortName(), (new \ReflectionClass($itemClass))->getShortName()), $info->fieldNodes, $info->path);
+            throw new \UnexpectedValueException(sprintf($errorMessage, (new \ReflectionClass($resourceClass))->getShortName(), (new \ReflectionClass($itemClass))->getShortName()));
         }
 
         return $resourceClass;